Add files via upload
Offline Analyses of Brain/Neural Exoskeleton control.
Control Accuracy (CA) based on brain and ocular signals was evaluated for data collected during the last run of SMR-ERD and HOVs calibration. Thus, CA evaluates the success of the calibration and infers the ability of participants to control the B/NE system. As established in the BCI field, CA was defined as ratio between correctly classified trials and the total number of trials. We defined a correctly performed EEG trial as one in which participants were able to downmodulate neural activity during movement trials, i.e., a negative modulation below the 75th percentile of SMR activity during the motor task, but not during rest periods. CA was evaluated based on the last 4 seconds of each 5 seconds-long activation interval, and the central 4 seconds of an inter-trial interval at which participants were at rest. A successful EOG trial was defined as a modulation above the 75th percentile of the ocular signal. In analogy to SMR-ERD, also the EOG CA was computed as the ratio between correctly classified and total number of trials. 
Time to Initialize (TTI) was assessed, i.e. participants’ time to react to cues and voluntarily modulate brain and ocular signals during calibration, as an estimate of bio-signal control intuitiveness. Intuitive control was defined as the initiation of biosignal modulation, i.e., reaching 50% of the maximal ERD or HOV, within less than 3 seconds after cue presentation.